requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
chromadb
unstructured
pypdf
tiktoken